feat: add {countries|landmarkTypes}ForFiltering methods
These methods in the Country and LandmarkType models return an array of
all countries/types with an "All" country/type appended. This list is
meant to be used only for filtering landmarks by country/type.

Previously I had added a literal "All" country/type record to the
database, which isn't really correct.