wip: commit working export implementation minus CSRF
The current implementation works as desired (Clicking "Export as JSON")
button saves displayed landmarks to disk and then makes the client
browser prompt for download.
The problem is that I am hardcoding the XSRF token in the POST request
to `landmarks.download`, and I need a way to automate this.